app/requirements.txt
functions-framework
google-cloud-storage
google-cloud-documentai
google-api-core
python-dotenv
google-cloud-aiplatform

# excel_updater dependencies
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas